Toggle antialias when switching to full screen
Some people - like me - use Terminus, the most awesome terminal font on
earth. But it only looks good with antialias disabled. When switching to
full screen and using a different font it makes sense to turn on
antialias - for the Cousine font - but when returning back it should go
back.